Migrate `Presentation` types to `packages/particles`.
