+++ packages/particles-ui-demo/src/state/list.ts
@@ -4,14 +4,15 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import type { Particles, Types } from "@breadboard-ai/particles-ui";
+import { Presentation } from "@breadboard-ai/particles";
+import type { Types } from "@breadboard-ai/particles-ui";
 import { SignalMap } from "signal-utils/map";
 
 export { List };
 
 class List implements Types.ItemList {
   items = new SignalMap<string, Types.ItemState>();
-  presentation: Particles.Presentation = {
+  presentation: Presentation = {
     type: "list",
     orientation: "vertical",
     behaviors: [],
