+++ packages/particles-ui-demo/src/ui/elements/list.ts
@@ -17,6 +17,7 @@ import "./button.js";
 import "./card.js";
 import "./hero-image.js";
 import "./segment.js";
+import { Orientation } from "@breadboard-ai/particles";
 
 @customElement("ui-list")
 export class UIList extends SignalWatcher(LitElement) {
@@ -24,7 +25,7 @@ export class UIList extends SignalWatcher(LitElement) {
   accessor list: ParticlesUI.Types.ItemList | null = null;
 
   @property({ reflect: true, type: String })
-  accessor orientation: ParticlesUI.Particles.Orientation = "vertical";
+  accessor orientation: Orientation = "vertical";
 
   @consume({ context: themeContext })
   accessor theme: UITheme | undefined;
