+++ packages/particles-ui-demo/src/ui/elements/segment.ts
@@ -13,17 +13,18 @@ import { themeContext } from "../context/theme";
 import { UITheme } from "../theme/theme";
 
 import * as ParticlesUI from "@breadboard-ai/particles-ui";
+import { Field, Orientation } from "@breadboard-ai/particles";
 
 @customElement("ui-segment")
 export class UISegment extends SignalWatcher(LitElement) {
   @property()
-  accessor fields: Record<string, ParticlesUI.Particles.Field> | null = null;
+  accessor fields: Record<string, Field> | null = null;
 
   @property()
   accessor values: Record<string, unknown> | null = null;
 
   @property()
-  accessor containerOrientation: ParticlesUI.Particles.Orientation = "vertical";
+  accessor containerOrientation: Orientation = "vertical";
 
   @property({ reflect: true, type: Boolean })
   accessor disabled = false;
@@ -42,7 +43,7 @@ export class UISegment extends SignalWatcher(LitElement) {
 
   #renderField(
     fieldName: string,
-    field: ParticlesUI.Particles.Field,
+    field: Field,
     value: unknown,
     theme: UITheme
   ) {
