+++ packages/particles-ui/src/types/types.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { type Presentation } from "./particles.js";
+import type { Presentation } from "@breadboard-ai/particles";
 
 export type ItemData = Record<string, string | boolean | Date>;
 
@@ -16,4 +16,19 @@ export type ItemState = {
 export type ItemList = {
   items: Map<string, ItemState>;
   presentation: Presentation;
+};
+
+export type TodoItemListTitle = string;
+
+/**
+ * The Receiver side of the channel, a proxy that represents the Generator.
+ */
+export type GeneratorProxy = {
+  requestAddItem(): Promise<void>;
+  requestUpdateField(
+    parentId: string,
+    id: string,
+    value: string | boolean
+  ): Promise<void>;
+  requestDelete(itemId: string): Promise<void>;
 };
